# Allow float stoichiometry