There is a decision to be made on the default conventions. 

Prior to this PR we always scaled ODE and jump rates by a combinatorial factor, treating the rate constants as microscopic rate constants. In this way our ODE models are the mean-field of our jump models.

However, such a scaling doesn't seem to me to make sense when having non-integer stoichiometry coefficients (unless we want to restrict non-integer stoichiometry to products and not allow for reactants, or scale by a gamma function which seems weird). 

In this PR I set the conventions as:
1. All the reactant and product stoichiometry of a given reaction are promoted to the same type.
2. If this type is an integer we rescale by default as we currently do.
3. If this type is not integer we error if the user does not explicitly turn off rescaling.

We could alternatively just rescale silently for the rate law of an individual reaction that involves non-integer stoichiometry. The problem is then users will get a mix of rescaled and not rescaled rate laws, which seems worse than giving them a useful error and forcing them to understand that they must turn off rescaling.

edit: This will also only work for ODEs currently, but I think that is fine as it is unclear to me what a non-integer stoichiometry even means in the stochastic case (i.e. what are the jump model's rate laws).