@isaacsas 2022-02-21
Note the promotion of type process probably will need to be updated to handle non-numeric types (i.e. Symbolics expressions). I'll save that for a future PR though.