@isaacsas 2022-02-21
@ChrisRackauckas this actually seems to generate reasonable ODEs in limited local testing, so if you agree with the conventions I choose I'll add some tests next.