@isaacsas 2022-02-21
Since earlier I had been worried this would have performance implications I `@benchmark`'ed before and after this PR with loading BCR and converting it to ODEs, I find:
```julia
prnbng = loadrxnetwork(BNGNetwork(),fname)
```
went from a median time of 436.518 ms to 437.124 ms
```julia
bosys = convert(ODESystem, rnbng)
```
from 2.151 s to 2.126 s
```julia
boprob = ODEProblem(bosys, Float64[], (0.,tf), Float64[])
```
from 1.028 s to 1.027 s.

So this doesn't look to have had any impact on the time to load and generate an ODE model.